+++ src/coordinator/coordinator.go
@@ -55,7 +55,7 @@ func NewCoordinatorImpl(datastore datastore.Datastore, raftServer ClusterConsens
 
 // Distributes the query across the cluster and combines the results. Yields as they come in ensuring proper order.
 // TODO: make this work even if there is a downed server in the cluster
-func (self *CoordinatorImpl) DistributeQuery(user common.User, db string, query *parser.Query, yield func(*protocol.Series) error) error {
+func (self *CoordinatorImpl) DistributeQuery(user common.User, db string, query *parser.SelectQuery, yield func(*protocol.Series) error) error {
 	if self.clusterConfiguration.IsSingleServer() {
 		return self.datastore.ExecuteQuery(user, db, query, yield, nil)
 	}
@@ -433,6 +433,22 @@ func (self *CoordinatorImpl) WriteSeriesData(user common.User, db string, series
 		}
 	}
 	return nil
+}
+
+func (self *CoordinatorImpl) DeleteSeriesData(user common.User, db string, query *parser.DeleteQuery) error {
+	if !user.HasWriteAccess(db) {
+		return common.NewAuthorizationError("Insufficient permission to write to %s", db)
+	}
+
+	if self.clusterConfiguration.IsSingleServer() {
+		return self.deleteSeriesDataLocally(user, db, query)
+	}
+
+	return fmt.Errorf("not implemented yet")
+}
+
+func (self *CoordinatorImpl) deleteSeriesDataLocally(user common.User, database string, query *parser.DeleteQuery) error {
+	return self.datastore.DeleteSeriesData(user, database, query)
 }
 
 func (self *CoordinatorImpl) writeSeriesToLocalStore(db *string, series *protocol.Series) error {