+++ src/datastore/leveldb_datastore.go
@@ -260,6 +260,27 @@ func (self *LevelDbDatastore) bytesToCurrentNumber(numberBytes []byte) uint64 {
 	return currentNumber
 }
 
+func (self *LevelDbDatastore) DeleteSeriesData(user common.User, database string, query *parser.DeleteQuery) error {
+	series := query.GetFromClause()
+	if series.Type != parser.FromClauseArray {
+		return fmt.Errorf("Merge and Inner joins can't be used with a delete query", series.Type)
+	}
+
+	for _, name := range series.Names {
+		var err error
+		if regex, ok := name.Name.GetCompiledRegex(); ok {
+			err = self.DeleteRangeOfRegex(user, database, regex, query.GetStartTime(), query.GetEndTime())
+		} else {
+			err = self.DeleteRangeOfSeries(database, name.Name.Name, query.GetStartTime(), query.GetEndTime())
+		}
+
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func (self *LevelDbDatastore) WriteSeriesData(database string, series *protocol.Series) error {
 	wb := levigo.NewWriteBatch()
 	defer wb.Close()
@@ -486,7 +507,7 @@ func (self *LevelDbDatastore) LogRequestAndAssignSequenceNumber(request *protoco
 }
 
 func (self *LevelDbDatastore) ExecuteQuery(user common.User, database string,
-	query *parser.Query, yield func(*protocol.Series) error,
+	query *parser.SelectQuery, yield func(*protocol.Series) error,
 	ringFilter func(database, series *string, time *int64) bool) error {
 
 	seriesAndColumns := query.GetReferencedColumns()
@@ -651,7 +672,7 @@ func isPointInRange(fieldId, startTime, endTime, point []byte) bool {
 }
 
 func (self *LevelDbDatastore) executeQueryForSeries(database, series string, columns []string,
-	query *parser.Query, yield func(*protocol.Series) error,
+	query *parser.SelectQuery, yield func(*protocol.Series) error,
 	ringFilter func(database, series *string, time *int64) bool) error {
 
 	startTimeBytes, endTimeBytes := self.byteArraysForStartAndEndTimes(common.TimeToMicroseconds(query.GetStartTime()), common.TimeToMicroseconds(query.GetEndTime()))
@@ -801,7 +822,7 @@ func (self *LevelDbDatastore) executeQueryForSeries(database, series string, col
 
 // Return the number of dropped ticks from filtering. if the series
 // had more than one alias, returns the min of all dropped ticks
-func (self *LevelDbDatastore) sendBatch(query *parser.Query, series *protocol.Series, yield func(series *protocol.Series) error) (int, error) {
+func (self *LevelDbDatastore) sendBatch(query *parser.SelectQuery, series *protocol.Series, yield func(series *protocol.Series) error) (int, error) {
 	dropped := int(math.MaxInt32)
 
 	for _, alias := range query.GetTableAliases(*series.Name) {