+++ src/parser/frees.c
@@ -71,12 +71,8 @@ free_error (error *error)
 }
 
 void
-close_query (query *q)
+free_select_query (select_query *q)
 {
-   if (q->error) {
-    free_error(q->error);
-   }
-
   if (q->c) {
     // free the columns
     free_value_array(q->c);
@@ -93,5 +89,36 @@ close_query (query *q)
   if (q->from_clause) {
     // free the from clause
     free_from_clause(q->from_clause);
+  }
+}
+
+void
+free_delete_query (delete_query *q)
+{
+  if (q->where_condition) {
+    free_condition(q->where_condition);
+  }
+
+  if (q->from_clause) {
+    // free the from clause
+    free_from_clause(q->from_clause);
+  }
+}
+
+void
+close_query (query *q)
+{
+   if (q->error) {
+    free_error(q->error);
+   }
+
+  if (q->select_query) {
+    free_select_query(q->select_query);
+    free(q->select_query);
+  }
+
+  if (q->delete_query) {
+    free_delete_query(q->delete_query);
+    free(q->delete_query);
   }
 }