+++ src/parser/parser_test.go
@@ -25,7 +25,7 @@ func ToValueArray(strings ...string) (values []*Value) {
 
 func (self *QueryParserSuite) TestInvalidFromClause(c *C) {
 	query := "select value from _t"
-	_, err := ParseQuery(query)
+	_, err := ParseSelectQuery(query)
 
 	c.Assert(err, ErrorMatches, ".*\\$undefined.*")
 }
@@ -36,7 +36,7 @@ func (self *QueryParserSuite) TestParseBasicSelectQuery(c *C) {
 		// semicolon is optional
 		"select value from t where c = '5'",
 	} {
-		q, err := ParseQuery(query)
+		q, err := ParseSelectQuery(query)
 		c.Assert(err, IsNil)
 
 		c.Assert(q.GetQueryString(), Equals, query)
@@ -56,8 +56,29 @@ func (self *QueryParserSuite) TestParseBasicSelectQuery(c *C) {
 	}
 }
 
+func (self *QueryParserSuite) TestParseDeleteQuery(c *C) {
+	query := "delete from foo where time > '2012-08-13' and time < '2013-08-13'"
+	queries, err := ParseQuery(query)
+	c.Assert(err, IsNil)
+
+	c.Assert(queries, HasLen, 1)
+
+	_q := queries[0]
+
+	c.Assert(_q.DeleteQuery, NotNil)
+
+	q := _q.DeleteQuery
+
+	c.Assert(q.GetQueryString(), Equals, query)
+
+	startTime, _ := time.Parse("2006-01-02", "2012-08-13")
+	endTime, _ := time.Parse("2006-01-02", "2013-08-13")
+	c.Assert(q.GetStartTime(), Equals, startTime)
+	c.Assert(q.GetEndTime(), Equals, endTime)
+}
+
 func (self *QueryParserSuite) TestSimpleFromClause(c *C) {
-	q, err := ParseQuery("select value from t;")
+	q, err := ParseSelectQuery("select value from t;")
 	c.Assert(err, IsNil)
 
 	fromClause := q.GetFromClause()
@@ -67,7 +88,7 @@ func (self *QueryParserSuite) TestSimpleFromClause(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseFromWithMergedTable(c *C) {
-	q, err := ParseQuery("select count(*) from newsletter.signups merge user.signups where time>now()-1d;")
+	q, err := ParseSelectQuery("select count(*) from newsletter.signups merge user.signups where time>now()-1d;")
 	c.Assert(err, IsNil)
 	fromClause := q.GetFromClause()
 	c.Assert(fromClause.Type, Equals, FromClauseMerge)
@@ -77,7 +98,7 @@ func (self *QueryParserSuite) TestParseFromWithMergedTable(c *C) {
 }
 
 func (self *QueryParserSuite) TestMultipleAggregateFunctions(c *C) {
-	q, err := ParseQuery("select first(bar), last(bar) from foo")
+	q, err := ParseSelectQuery("select first(bar), last(bar) from foo")
 	c.Assert(err, IsNil)
 	columns := q.GetColumnNames()
 	c.Assert(columns, HasLen, 2)
@@ -86,7 +107,7 @@ func (self *QueryParserSuite) TestMultipleAggregateFunctions(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseFromWithJoinedTable(c *C) {
-	q, err := ParseQuery("select max(newsletter.signups.value, user.signups.value) from newsletter.signups inner join user.signups where time>now()-1d;")
+	q, err := ParseSelectQuery("select max(newsletter.signups.value, user.signups.value) from newsletter.signups inner join user.signups where time>now()-1d;")
 	c.Assert(err, IsNil)
 	fromClause := q.GetFromClause()
 	c.Assert(fromClause.Type, Equals, FromClauseInnerJoin)
@@ -96,7 +117,7 @@ func (self *QueryParserSuite) TestParseFromWithJoinedTable(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithInsensitiveRegexTables(c *C) {
-	q, err := ParseQuery("select email from /users.*/i where time>now()-2d;")
+	q, err := ParseSelectQuery("select email from /users.*/i where time>now()-2d;")
 	c.Assert(err, IsNil)
 
 	fromClause := q.GetFromClause()
@@ -108,7 +129,7 @@ func (self *QueryParserSuite) TestParseSelectWithInsensitiveRegexTables(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithRegexTables(c *C) {
-	q, err := ParseQuery("select email from /users.*/ where time>now()-2d;")
+	q, err := ParseSelectQuery("select email from /users.*/ where time>now()-2d;")
 	c.Assert(err, IsNil)
 
 	fromClause := q.GetFromClause()
@@ -121,7 +142,7 @@ func (self *QueryParserSuite) TestParseSelectWithRegexTables(c *C) {
 }
 
 func (self *QueryParserSuite) TestMergeFromClause(c *C) {
-	q, err := ParseQuery("select value from t1 merge t2 where c = '5';")
+	q, err := ParseSelectQuery("select value from t1 merge t2 where c = '5';")
 	c.Assert(err, IsNil)
 
 	c.Assert(q.GetColumnNames(), DeepEquals, ToValueArray("value"))
@@ -139,14 +160,14 @@ func (self *QueryParserSuite) TestMergeFromClause(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithoutWhereClause(c *C) {
-	q, err := ParseQuery("select value, time from t;")
+	q, err := ParseSelectQuery("select value, time from t;")
 	c.Assert(err, IsNil)
 	c.Assert(q.GetColumnNames(), DeepEquals, ToValueArray("value", "time"))
 	c.Assert(q.GetWhereCondition(), IsNil)
 }
 
 func (self *QueryParserSuite) TestParseSelectWithUpperCase(c *C) {
-	q, err := ParseQuery("SELECT VALUE, TIME FROM t WHERE C = '5';")
+	q, err := ParseSelectQuery("SELECT VALUE, TIME FROM t WHERE C = '5';")
 	c.Assert(err, IsNil)
 	c.Assert(q.GetColumnNames(), DeepEquals, ToValueArray("VALUE", "TIME"))
 	w := q.GetWhereCondition()
@@ -161,7 +182,7 @@ func (self *QueryParserSuite) TestParseSelectWithUpperCase(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithMultipleColumns(c *C) {
-	q, err := ParseQuery("select value, time from t;")
+	q, err := ParseSelectQuery("select value, time from t;")
 
 	c.Assert(err, IsNil)
 	columns := q.GetColumnNames()
@@ -171,7 +192,7 @@ func (self *QueryParserSuite) TestParseSelectWithMultipleColumns(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithInequality(c *C) {
-	q, err := ParseQuery("select value, time from t where c < 5;")
+	q, err := ParseSelectQuery("select value, time from t where c < 5;")
 	c.Assert(err, IsNil)
 	w := q.GetWhereCondition()
 
@@ -186,7 +207,7 @@ func (self *QueryParserSuite) TestParseSelectWithInequality(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithTimeCondition(c *C) {
-	q, err := ParseQuery("select value, time from t where time > now() - 1d;")
+	q, err := ParseSelectQuery("select value, time from t where time > now() - 1d;")
 	c.Assert(err, IsNil)
 
 	// note: the time condition will be removed
@@ -206,7 +227,7 @@ func (self *QueryParserSuite) TestParseSelectWithPartialTimeString(c *C) {
 	} {
 		t, err := time.Parse("2006-01-02 15:04:05", expected)
 		c.Assert(err, IsNil)
-		q, err := ParseQuery(fmt.Sprintf("select value, time from t where time > '%s';", actual))
+		q, err := ParseSelectQuery(fmt.Sprintf("select value, time from t where time > '%s';", actual))
 		c.Assert(err, IsNil)
 
 		// note: the time condition will be removed
@@ -220,7 +241,7 @@ func (self *QueryParserSuite) TestParseSelectWithPartialTimeString(c *C) {
 func (self *QueryParserSuite) TestParseSelectWithTimeString(c *C) {
 	t, err := time.Parse("2006-01-02 15:04:05", "2013-08-12 23:32:01.232")
 	c.Assert(err, IsNil)
-	q, err := ParseQuery("select value, time from t where time > '2013-08-12 23:32:01.232';")
+	q, err := ParseSelectQuery("select value, time from t where time > '2013-08-12 23:32:01.232';")
 	c.Assert(err, IsNil)
 
 	// note: the time condition will be removed
@@ -234,7 +255,7 @@ func (self *QueryParserSuite) TestParseSelectWithTimeString(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithAnd(c *C) {
-	q, err := ParseQuery("select value from cpu.idle where value > exp() * 2 and value < exp() * 3;")
+	q, err := ParseSelectQuery("select value from cpu.idle where value > exp() * 2 and value < exp() * 3;")
 	c.Assert(err, IsNil)
 
 	w := q.GetWhereCondition()
@@ -265,7 +286,7 @@ func (self *QueryParserSuite) TestParseSelectWithAnd(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithGroupBy(c *C) {
-	q, err := ParseQuery("select count(*) from users.events group by user_email,time(1h) where time>now()-1d;")
+	q, err := ParseSelectQuery("select count(*) from users.events group by user_email,time(1h) where time>now()-1d;")
 	c.Assert(err, IsNil)
 
 	c.Assert(q.GetColumnNames(), HasLen, 1)
@@ -288,7 +309,7 @@ func (self *QueryParserSuite) TestParseSelectWithGroupBy(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseFromWithNestedFunctions(c *C) {
-	q, err := ParseQuery("select top(10, count(*)) from users.events;")
+	q, err := ParseSelectQuery("select top(10, count(*)) from users.events;")
 	c.Assert(err, IsNil)
 	c.Assert(q.GetColumnNames(), HasLen, 1)
 	column := q.GetColumnNames()[0]
@@ -305,7 +326,7 @@ func (self *QueryParserSuite) TestParseFromWithNestedFunctions(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseWhereClausePrecedence(c *C) {
-	q, err := ParseQuery("select value from cpu.idle where value > 90 and other_value > 10.0 or value > 80 and other_value > 20;")
+	q, err := ParseSelectQuery("select value from cpu.idle where value > 90 and other_value > 10.0 or value > 80 and other_value > 20;")
 	c.Assert(err, IsNil)
 
 	whereCondition := q.GetWhereCondition()
@@ -331,7 +352,7 @@ func (self *QueryParserSuite) TestParseWhereClausePrecedence(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseWhereClauseParentheses(c *C) {
-	q, err := ParseQuery("select value from cpu.idle where value > 90 and (other_value > 10 or value > 80) and other_value > 20;")
+	q, err := ParseSelectQuery("select value from cpu.idle where value > 90 and (other_value > 10 or value > 80) and other_value > 20;")
 	c.Assert(err, IsNil)
 
 	whereCondition := q.GetWhereCondition()
@@ -346,16 +367,16 @@ func (self *QueryParserSuite) TestParseWhereClauseParentheses(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithOrderByAndLimit(c *C) {
-	q, err := ParseQuery("select value from t order asc limit 10;")
+	q, err := ParseSelectQuery("select value from t order asc limit 10;")
 	c.Assert(err, IsNil)
 	c.Assert(q.Limit, Equals, 10)
 	c.Assert(q.Ascending, Equals, true)
 
-	q, err = ParseQuery("select value from t order desc;")
+	q, err = ParseSelectQuery("select value from t order desc;")
 	c.Assert(err, IsNil)
 	c.Assert(q.Ascending, Equals, false)
 
-	q, err = ParseQuery("select value from t limit 20;")
+	q, err = ParseSelectQuery("select value from t limit 20;")
 	c.Assert(err, IsNil)
 	c.Assert(q.Limit, Equals, 20)
 	// descending is the default
@@ -363,7 +384,7 @@ func (self *QueryParserSuite) TestParseSelectWithOrderByAndLimit(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseFromWithNestedFunctions2(c *C) {
-	q, err := ParseQuery("select count(distinct(email)) from user.events where time>now()-1d group by time(15m);")
+	q, err := ParseSelectQuery("select count(distinct(email)) from user.events where time>now()-1d group by time(15m);")
 	c.Assert(err, IsNil)
 	c.Assert(q.GetColumnNames(), HasLen, 1)
 	column := q.GetColumnNames()[0]
@@ -384,12 +405,12 @@ func (self *QueryParserSuite) TestParseFromWithNestedFunctions2(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithInvalidRegex(c *C) {
-	_, err := ParseQuery("select email from users.events where email =~ /[/i and time>now()-2d;")
+	_, err := ParseSelectQuery("select email from users.events where email =~ /[/i and time>now()-2d;")
 	c.Assert(err, ErrorMatches, ".*missing closing.*")
 }
 
 func (self *QueryParserSuite) TestParseSelectWithRegexCondition(c *C) {
-	q, err := ParseQuery("select email from users.events where email =~ /gmail\\.com/i and time>now()-2d;")
+	q, err := ParseSelectQuery("select email from users.events where email =~ /gmail\\.com/i and time>now()-2d;")
 	c.Assert(err, IsNil)
 	w := q.GetWhereCondition()
 
@@ -403,7 +424,7 @@ func (self *QueryParserSuite) TestParseSelectWithRegexCondition(c *C) {
 }
 
 func (self *QueryParserSuite) TestParseSelectWithComplexArithmeticOperations(c *C) {
-	q, err := ParseQuery("select value from cpu.idle where .30 < value * 1 / 3 ;")
+	q, err := ParseSelectQuery("select value from cpu.idle where .30 < value * 1 / 3 ;")
 	c.Assert(err, IsNil)
 
 	boolExpression, ok := q.GetWhereCondition().GetBoolExpression()
@@ -435,7 +456,7 @@ func (self *QueryParserSuite) TestTimeConditionWithFloats(c *C) {
 		fmt.Sprintf("select * from foo where time > %ds", startTime),
 		fmt.Sprintf("select * from foo where time > %d.0s", startTime),
 	} {
-		q, err := ParseQuery(query)
+		q, err := ParseSelectQuery(query)
 		c.Assert(err, IsNil)
 		c.Assert(q.GetStartTime(), Equals, time.Unix(startTime, 0).UTC())
 	}
@@ -443,7 +464,7 @@ func (self *QueryParserSuite) TestTimeConditionWithFloats(c *C) {
 
 func (self *QueryParserSuite) TestQureyWithInCondition(c *C) {
 	query := "select * from foo where bar in ('baz', 'bazz')"
-	q, err := ParseQuery(query)
+	q, err := ParseSelectQuery(query)
 	c.Assert(err, IsNil)
 	condition := q.GetWhereCondition()
 	expr, ok := condition.GetBoolExpression()