+++ CHANGELOG.md
@@ -152,6 +152,7 @@
 - [Issue #102](https://github.com/influxdb/influxdb/issues/102). Support expressions in where condition
 - [Issue #101](https://github.com/influxdb/influxdb/issues/101). Support expressions in aggregates
 - [Issue #62](https://github.com/influxdb/influxdb/issues/62). Support updating and deleting column values
+- Support delete queries
 
 ## Bugfixes